+++ packages/shared/src/components/cards/Leaderboard/LeaderboardList.tsx
@@ -4,6 +4,11 @@ import { LeaderboardListContainerProps } from './common';
 import classed from '../../../lib/classed';
 import { ElementPlaceholder } from '../../ElementPlaceholder';
 
+export interface CommonLeaderboardProps<T extends Iterable<unknown>>
+  extends Omit<LeaderboardListProps, 'children'> {
+  items: T;
+}
+
 export interface LeaderboardListProps {
   containerProps: Omit<LeaderboardListContainerProps, 'children'>;
   isLoading: boolean;
