+++ packages/shared/src/components/cards/Leaderboard/index.ts
@@ -1,5 +1,7 @@
 export * from './LeaderboardList';
 export * from './LeaderboardListItem';
 export * from './LeaderboardListContainer';
+export * from './SourceTopList';
+export * from './UserTopList';
+export * from './TagTopList';
 export * from './common';
-export * from './lists';
